mess: stop using `eval`

+++ users/somasis/mess.nix
@@ -1,59 +1,43 @@
 # mess - Set up `mess` integration.
 { config, pkgs, ... }:
 let
-  mess = pkgs.writeShellScriptBin "mess" ''
-    usage() {
-        printf 'usage: %s [-s]\n' "''${0##*/}" >&2
-        exit 69
-    }
+  mess = pkgs.writeShellApplication {
+    name = "mess";
 
-    print_sh() {
-        cat <<'EOF'
-    mess() {
-        if [ $# -eq 0 ]; then
-            cd "$(command mess)" && return
-        else
-            command mess "$@"
-        fi
-    }
-    EOF
-        exit 0
-    }
+    runtimeInputs = [ pkgs.coreutils ];
 
-    MESSDIR=''${MESSDIR:-~/mess}
+    text = ''
+      set -eu
+      set -o pipefail
 
-    while getopts :s arg >/dev/null 2>&1; do
-        case "''${arg}" in
-            s)
-                print_sh
-                ;;
-            ?)
-                printf "unknown argument -- %s\n" "''${OPTARG}"
-                usage
-                ;;
-        esac
-    done
+      usage() {
+          printf 'usage: %s\n' "''${0##*/}" >&2
+          exit 69
+      }
 
-    # Avoid weeks 00 and 53 in the week number, make it more like natural language.
-    week=$(date +%W)
-    case "''${week}" in 00) week=01 ;; 53) week=52 ;; esac
+      : "''${MESSDIR:=$HOME/mess}"
 
-    current=$(date +%Y/"''${week}")
+      [ "$#" -gt 0 ] && usage
 
-    if ! [ -d "''${MESSDIR}/''${current}" ]; then
-        mkdir -p "''${MESSDIR}/''${current}"
-        mkdir -p "''${MESSDIR}/''${current}/incoming" "''${MESSDIR}/''${current}/src"
-        printf "Created messdir '%s'.\n" "''${current}" >&2
-    fi
+      # Avoid weeks 00 and 53 in the week number, make it more like natural language.
+      week=$(date +%W)
+      case "''${week}" in 00) week=01 ;; 53) week=52 ;; esac
 
-    [ "$(readlink -f "''${MESSDIR}/current")" = "''${MESSDIR}/''${current}" ] \
-        || {
-             rm -f "''${MESSDIR}"/current
-             ln -sf "''${current}" "''${MESSDIR}/current"
-        }
+      current=$(date +%Y/"''${week}")
 
-    printf '%s\n' "''${MESSDIR}/current"
-  '';
+      if ! [ -d "''${MESSDIR}/''${current}" ]; then
+          mkdir -p "''${MESSDIR}/''${current}"
+          printf "Created mess '%s'.\n" "''${current}" >&2
+      fi
+
+      if ! [ "$(readlink -f "''${MESSDIR}/current")" = "''${MESSDIR}/''${current}" ]; then
+          rm -f "''${MESSDIR}"/current
+          ln -sf "''${current}" "''${MESSDIR}/current"
+      fi
+
+      printf '%s\n' "''${MESSDIR}/current"
+    '';
+  };
 in
 {
   home.packages = [ mess ];
@@ -98,9 +82,13 @@ in
   };
 
   programs.bash.initExtra = ''
-    eval "$(mess -s)"
-
-    mkdir -p ~/mess/current/src >/dev/null 2>&1
+    mess() {
+        if [ "$#" -eq 0 ]; then
+            cd "$(command mess)"
+        else
+            command mess "$@"
+        fi
+    }
 
     src() {
         CDPATH="$HOME/mess/current/src:$HOME/src:$HOME/src/nix:$HOME/src/discord" cd "''${@:-}"
